refactor(issuance): consolidate types with existing session module, address PR reviews
Replaced duplicate Session/SessionState/TxCodeSpec in domain/models.rs
with the existing IssuanceSession/IssuanceState from session/mod.rs.
TxCode validation now uses the spec directly from
offer.grants.pre_authorized_code.tx_code instead of a wrapper struct.
Moved ProcessingStep, FailureStep, TxCodeValidationError into the
session module. Split issuance handler into tx_code.rs and cancel.rs
submodules. Simplified SSE to only Processing and Failed variants.
Restored AppState and doc comments, used nested routes.

Hi @Hermann-Core — after pulling main I saw the existing session
module and understood your comments about duplication and scope creep.
All types now reuse what's already in session:: instead of reinventing
them in domain/models.rs.
